+++ src/main/java/net/minecraft/entity/item/EntityBoat.java
@@ -423,7 +423,7 @@ public class EntityBoat extends Entity {
     }
 
     public void setDamageTaken(float p_70266_1_) {
-        this.dataWatcher.updateObject(19, Float.valueOf(p_70266_1_));
+        this.dataWatcher.updateObject(19, p_70266_1_);
     }
 
     public float getDamageTaken() {
@@ -431,7 +431,7 @@ public class EntityBoat extends Entity {
     }
 
     public void setTimeSinceHit(int p_70265_1_) {
-        this.dataWatcher.updateObject(17, Integer.valueOf(p_70265_1_));
+        this.dataWatcher.updateObject(17, p_70265_1_);
     }
 
     public int getTimeSinceHit() {
@@ -439,7 +439,7 @@ public class EntityBoat extends Entity {
     }
 
     public void setForwardDirection(int p_70269_1_) {
-        this.dataWatcher.updateObject(18, Integer.valueOf(p_70269_1_));
+        this.dataWatcher.updateObject(18, p_70269_1_);
     }
 
     public int getForwardDirection() {
