+++ src/main/java/net/minecraft/entity/item/EntityMinecart.java
@@ -87,7 +87,7 @@ public abstract class EntityMinecart extends Entity implements IWorldNameable {
         this.dataWatcher.addObject(19, 0.0F);
         this.dataWatcher.addObject(20, 0);
         this.dataWatcher.addObject(21, 6);
-        this.dataWatcher.addObject(22, Byte.valueOf((byte) 0));
+        this.dataWatcher.addObject(22, (byte) 0);
     }
 
     public AxisAlignedBB getCollisionBox(Entity entityIn) {
@@ -767,7 +767,7 @@ public abstract class EntityMinecart extends Entity implements IWorldNameable {
     }
 
     public void setDamage(float p_70492_1_) {
-        this.dataWatcher.updateObject(19, Float.valueOf(p_70492_1_));
+        this.dataWatcher.updateObject(19, p_70492_1_);
     }
 
     public float getDamage() {
@@ -775,7 +775,7 @@ public abstract class EntityMinecart extends Entity implements IWorldNameable {
     }
 
     public void setRollingAmplitude(int p_70497_1_) {
-        this.dataWatcher.updateObject(17, Integer.valueOf(p_70497_1_));
+        this.dataWatcher.updateObject(17, p_70497_1_);
     }
 
     public int getRollingAmplitude() {
@@ -783,7 +783,7 @@ public abstract class EntityMinecart extends Entity implements IWorldNameable {
     }
 
     public void setRollingDirection(int p_70494_1_) {
-        this.dataWatcher.updateObject(18, Integer.valueOf(p_70494_1_));
+        this.dataWatcher.updateObject(18, p_70494_1_);
     }
 
     public int getRollingDirection() {
@@ -809,12 +809,12 @@ public abstract class EntityMinecart extends Entity implements IWorldNameable {
     }
 
     public void func_174899_a(IBlockState p_174899_1_) {
-        this.getDataWatcher().updateObject(20, Integer.valueOf(Block.getStateId(p_174899_1_)));
+        this.getDataWatcher().updateObject(20, Block.getStateId(p_174899_1_));
         this.setHasDisplayTile(true);
     }
 
     public void setDisplayTileOffset(int p_94086_1_) {
-        this.getDataWatcher().updateObject(21, Integer.valueOf(p_94086_1_));
+        this.getDataWatcher().updateObject(21, p_94086_1_);
         this.setHasDisplayTile(true);
     }
 
@@ -823,7 +823,7 @@ public abstract class EntityMinecart extends Entity implements IWorldNameable {
     }
 
     public void setHasDisplayTile(boolean p_94096_1_) {
-        this.getDataWatcher().updateObject(22, Byte.valueOf((byte) (p_94096_1_ ? 1 : 0)));
+        this.getDataWatcher().updateObject(22, (byte) (p_94096_1_ ? 1 : 0));
     }
 
     public void setCustomNameTag(String name) {
@@ -883,13 +883,13 @@ public abstract class EntityMinecart extends Entity implements IWorldNameable {
         }
 
         public static EntityMinecart.EnumMinecartType byNetworkID(int id) {
-            EntityMinecart.EnumMinecartType entityminecart$enumminecarttype = ID_LOOKUP.get(Integer.valueOf(id));
+            EntityMinecart.EnumMinecartType entityminecart$enumminecarttype = ID_LOOKUP.get(id);
             return entityminecart$enumminecarttype == null ? RIDEABLE : entityminecart$enumminecarttype;
         }
 
         static {
             for (EntityMinecart.EnumMinecartType entityminecart$enumminecarttype : values()) {
-                ID_LOOKUP.put(Integer.valueOf(entityminecart$enumminecarttype.getNetworkID()), entityminecart$enumminecarttype);
+                ID_LOOKUP.put(entityminecart$enumminecarttype.getNetworkID(), entityminecart$enumminecarttype);
             }
         }
     }
