+++ src/main/java/net/minecraft/entity/monster/EntityEnderman.java
@@ -258,7 +258,7 @@ public class EntityEnderman extends EntityMob {
     }
 
     public void setHeldBlockState(IBlockState state) {
-        this.dataWatcher.updateObject(16, Short.valueOf((short) (Block.getStateId(state) & 65535)));
+        this.dataWatcher.updateObject(16, (short) (Block.getStateId(state) & 65535));
     }
 
     public IBlockState getHeldBlockState() {
@@ -310,7 +310,7 @@ public class EntityEnderman extends EntityMob {
     }
 
     public void setScreaming(boolean screaming) {
-        this.dataWatcher.updateObject(18, Byte.valueOf((byte) (screaming ? 1 : 0)));
+        this.dataWatcher.updateObject(18, (byte) (screaming ? 1 : 0));
     }
 
     static {
