+++ src/main/java/net/minecraft/entity/monster/EntitySkeleton.java
@@ -267,7 +267,7 @@ public class EntitySkeleton extends EntityMob implements IRangedAttackMob {
     }
 
     public void setSkeletonType(int p_82201_1_) {
-        this.dataWatcher.updateObject(13, Byte.valueOf((byte) p_82201_1_));
+        this.dataWatcher.updateObject(13, (byte) p_82201_1_);
         this.isImmuneToFire = p_82201_1_ == 1;
 
         if (p_82201_1_ == 1) {
