+++ src/main/java/net/minecraft/world/storage/MapStorage.java
@@ -124,7 +124,7 @@ public class MapStorage {
 
                     if (nbtbase instanceof NBTTagShort nbttagshort) {
                         short short1 = nbttagshort.getShort();
-                        this.idCounts.put(s, Short.valueOf(short1));
+                        this.idCounts.put(s, short1);
                     }
                 }
             }
@@ -137,9 +137,9 @@ public class MapStorage {
         Short oshort = this.idCounts.get(key);
 
         if (oshort == null) {
-            oshort = Short.valueOf((short) 0);
+            oshort = (short) 0;
         } else {
-            oshort = Short.valueOf((short) (oshort.shortValue() + 1));
+            oshort = (short) (oshort.shortValue() + 1);
         }
 
         this.idCounts.put(key, oshort);
