+++ src/main/java/net/optifine/shaders/Shaders.java
@@ -888,7 +888,7 @@ public class Shaders {
             String s = "/shaders/world" + i;
 
             if (shaderPack.hasDirectory(s)) {
-                shaderPackDimensions.add(Integer.valueOf(i));
+                shaderPackDimensions.add(i);
             }
         }
 
@@ -1597,7 +1597,7 @@ public class Shaders {
         int i = EXTFramebufferObject.glCheckFramebufferStatusEXT(36160);
 
         if (i != 36053) {
-            System.err.format("FramebufferStatus 0x%04X at %s\n", Integer.valueOf(i), location);
+            System.err.format("FramebufferStatus 0x%04X at %s\n", i, location);
         }
 
         return i;
@@ -1609,11 +1609,11 @@ public class Shaders {
         if (i != 0 && GlErrors.isEnabled(i)) {
             String s = Config.getGlErrorString(i);
             String s1 = getErrorInfo(i, location);
-            String s2 = String.format("OpenGL error: %s (%s)%s, at: %s", Integer.valueOf(i), s, s1, location);
+            String s2 = String.format("OpenGL error: %s (%s)%s, at: %s", i, s, s1, location);
             SMCLog.severe(s2);
 
             if (Config.isShowGlErrors() && TimedEvent.isActive("ShowGlErrorShaders", 10000L)) {
-                String s3 = I18n.format("of.message.openglError", Integer.valueOf(i), s);
+                String s3 = I18n.format("of.message.openglError", i, s);
                 printChat(s3);
             }
         }
@@ -1897,7 +1897,7 @@ public class Shaders {
             if (currentWorld != null) {
                 int i = currentWorld.provider.getDimensionId();
 
-                if (shaderPackDimensions.contains(Integer.valueOf(i))) {
+                if (shaderPackDimensions.contains(i)) {
                     s = "world" + i + "/";
                 }
             }
@@ -2526,7 +2526,7 @@ public class Shaders {
 
                                                                 if (vector4f != null) {
                                                                     gbuffersClearColor[j1] = vector4f;
-                                                                    SMCLog.info("%s clear color: %s %s %s %s", s3, Float.valueOf(vector4f.getX()), Float.valueOf(vector4f.getY()), Float.valueOf(vector4f.getZ()), Float.valueOf(vector4f.getW()));
+                                                                    SMCLog.info("%s clear color: %s %s %s %s", s3, vector4f.getX(), vector4f.getY(), vector4f.getZ(), vector4f.getW());
                                                                 } else {
                                                                     SMCLog.warning("Invalid color value: " + shaderline.getValue());
                                                                 }
@@ -2973,7 +2973,7 @@ public class Shaders {
             for (ResourceLocation resourcelocation : Block.blockRegistry.getKeys()) {
                 Block block = Block.blockRegistry.getObject(resourcelocation);
                 int i = Block.blockRegistry.getIDForObject(block);
-                mapBlockToEntityData.put(block, Integer.valueOf(i));
+                mapBlockToEntityData.put(block, i);
             }
         }
 
@@ -2998,7 +2998,7 @@ public class Shaders {
                         Block block1 = Block.getBlockFromName(s2);
 
                         if (block1 != null) {
-                            mapBlockToEntityData.put(block1, Integer.valueOf(j));
+                            mapBlockToEntityData.put(block1, j);
                         } else {
                             SMCLog.warning("Unknown block name %s", s2);
                         }
@@ -3481,8 +3481,8 @@ public class Shaders {
             int j = getDimensionId(world);
 
             if (j != i) {
-                boolean flag = shaderPackDimensions.contains(Integer.valueOf(i));
-                boolean flag1 = shaderPackDimensions.contains(Integer.valueOf(j));
+                boolean flag = shaderPackDimensions.contains(i);
+                boolean flag1 = shaderPackDimensions.contains(j);
 
                 if (flag || flag1) {
                     uninit();
