+++ src/main/java/net/optifine/util/ArrayUtils.java
@@ -134,7 +134,7 @@ public class ArrayUtils {
                     stringbuffer.append(separator);
                 }
 
-                stringbuffer.append(String.format(format, Float.valueOf(f)));
+                stringbuffer.append(String.format(format, f));
             }
 
             return stringbuffer.toString();
